Move the USB plugins to V3.

+++ plugins/romloader/usb/netx/src/usb_emsys/usb_io.c
@@ -60,11 +60,13 @@ void usb_io_write_fifo(unsigned int uiDwOffset, unsigned int uiByteCount, const 
 
 	uiByteCnt = 0;
 	ulValue = 0;
-	while( pulDc<pulDe ) {
+	while( pulDc<pulDe )
+	{
 		ulValue >>= 8;
-		ulValue  |= (*pucSc++)<<24;
+		ulValue  |= ((unsigned long)(*pucSc++)) << 24U;
 		uiByteCnt++;
-		if( (uiByteCnt&3)==0 ) {
+		if( (uiByteCnt&3U)==0U )
+		{
 			*pulDc++ = ulValue;
 		}
 	}
